+++ packages/safe-ds-lang/src/language/graphical-editor/ast-parser/argument.ts
@@ -1,10 +1,10 @@
 import { isSdsLiteral, SdsArgument } from '../../generated/ast.js';
-import { CustomError } from '../global.js';
 import { Call } from './call.js';
 import { Placeholder } from './placeholder.js';
 import { Expression, GenericExpression } from './expression.js';
 import { Parameter } from './parameter.js';
 import { Parser } from './parser.js';
+import { CustomError } from '../types.js';
 
 export class Argument {
     constructor(
